+++ src/schema/user.js
@@ -1,0 +1,16 @@
+import { gql } from "apollo-server-express";
+
+export default gql`
+  extend type Query {
+    me: User
+    user(id: ID!): User
+    users: [User!]
+    onlineUsers: [User!]!
+  }
+
+  type User {
+    id: ID!
+    username: String!
+    messages: [Message!]
+  }
+`;
